django_backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
firebase-admin==6.2.0
pyfcm==1.5.4
django-cors-headers==4.3.1
python-decouple==3.8
requests==2.31.0
django-otp==1.1.6
python-dotenv==3.0.0